refactor: Updated LoginView.xaml.cs to use new github authentication event arguments
