+++ BananaGit/BananaGit/EventArgExtensions/CredentialsEventArgs.cs
@@ -6,7 +6,7 @@
 public class CredentialsEventArgs(bool success) : System.EventArgs
 {
     /// <summary>
-    /// True if user succesfully logged in, false otherwise
+    /// True if user successfully logged in, false otherwise
     /// </summary>
     public bool LoginSuccess { get; set; } = success;
 }
