+++ BananaGit/BananaGit/Views/DialogueViews/LoginView.xaml
@@ -1,5 +1,5 @@
 ﻿<Window
-    x:Class="BananaGit.Views.LoginView"
+    x:Class="BananaGit.Views.DialogueViews.LoginView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -30,15 +30,6 @@
         <StackPanel Grid.Row="1">
 
 
-            <Label HorizontalAlignment="Center" Content="Username" />
-            <TextBox
-                Grid.Row="1"
-                Width="100"
-                HorizontalAlignment="Center"
-                VerticalAlignment="Center"
-                Background="Black"
-                Foreground="White"
-                Text="{Binding Username}" />
             <Label HorizontalAlignment="Center" Content="Email" />
             <TextBox
                 Grid.Row="1"
@@ -48,15 +39,7 @@
                 Background="Black"
                 Foreground="White"
                 Text="{Binding Email}" />
-            <Label HorizontalAlignment="Center" Content="Personal Token" />
-            <TextBox
-                Grid.Row="1"
-                Width="100"
-                HorizontalAlignment="Center"
-                VerticalAlignment="Center"
-                Background="Black"
-                Foreground="White"
-                Text="{Binding UserToken}" />
+
         </StackPanel>
 
         <Button
